chore: move constant fixtures
